+++ cw-db/src/test/java/edu/uob/DropDatabase.java
@@ -1,6 +1,7 @@
 package edu.uob;
 
 import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -13,6 +14,11 @@ public class DropDatabase {
     public void setup(){
         if(dbServer == null) dbServer = new DBServer();
     }
+    @BeforeAll
+    public static void makeFolder(){
+        File file = new File(Paths.get("databases").toUri());
+        if(!file.exists()) file.mkdir();
+    }
 
     String generateRandomName() {
         StringBuilder randomName = new StringBuilder();
@@ -23,8 +29,12 @@ public class DropDatabase {
     @AfterAll
     public static void cleanFolder(){
         File file = new File(Paths.get("databases").toAbsolutePath().toString());
+        File[] fileList = file.listFiles();
+        if(fileList == null) return;
         for(File internalDirectory : file.listFiles()){
             if(internalDirectory.getName().equalsIgnoreCase("testDB")) continue;
+            File[] internalFileList = internalDirectory.listFiles();
+            if(internalFileList == null) continue;
             for(File f : internalDirectory.listFiles()) f.delete();
             internalDirectory.delete();
         }
