+++ cw-db/src/test/java/edu/uob/ExampleDBTests.java
@@ -4,6 +4,7 @@ import edu.uob.Controller.IOController;
 import edu.uob.Model.Database;
 import edu.uob.Utils.Utils;
 import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import java.io.*;
@@ -24,11 +25,21 @@ public class ExampleDBTests {
         server = new DBServer();
     }
 
+    @BeforeAll
+    public static void makeFolder(){
+        File file = new File(Paths.get("databases").toUri());
+        if(!file.exists()) file.mkdir();
+    }
+
     @AfterAll
     public static void cleanFolder(){
         File file = new File(Paths.get("databases").toAbsolutePath().toString());
+        File[] fileList = file.listFiles();
+        if(fileList == null) return;
         for(File internalDirectory : file.listFiles()){
             if(internalDirectory.getName().equalsIgnoreCase("testDB")) continue;
+            File[] internalFileList = internalDirectory.listFiles();
+            if(internalFileList == null) continue;
             for(File f : internalDirectory.listFiles()) f.delete();
             internalDirectory.delete();
         }
